Working VR update
Added the VR component and prototyped the area to area movement based on enemy kills, this can all be changed and adjusted as we progress.  The movement in editor is set up for a PS4 controller for the time being, feel free to add keyboard and mouse support.  I will refine the functionality of the scripts going forward.